Fix datasets compilations
- Re-add accidentally dropped contexts
- Move docs out of syscore again
- Make new common datasets union
- Keep old datasets union for all of syscore, common, docs
- Fix duplicate description of docs dataset
